# provider/openapi.yaml
openapi: 3.0.0
info:
    title: Alias Provider backend API
    version: 0.1
components:
...
    schemas:
...
        user_meta:
            application/json:
                schema:
                    type: object
                    description: Any kind of metadata linked to the current user
paths:
...
    /api/session/:
        get:
            summary: Returns the current active session.
            tags:
                - Session
            responses:
                '401':
                    $ref: '#/components/requests/sessions/responses/401'

                '200':
                    description: Current session
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: "ok"
                                    publicKey:
                                        type: string
                                        format: byte
                                        description: user's public key
...
    /api/session/login:
        post:
            summary: Login an user, create an active session for the user.
            tags:
                - Session
            parameters:
                - name: token
                  in: anychain's signature
                  required: true
                  description: Anychain's signature with the user's private key of the latest seed returned by `/api/session/seed`.
                  schema:
                    type: string
                    format: json
            responses:
                '400':
                    description: Request is malformed. Signature is invalid or seed is unknown.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: "error"
                                    reason:
                                        type: string
                                        example: "bad proof: seed not valid"

                '200':
                    description: User is successfuly logged in.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: string
                                        example: "ok"
    /api/session/logout:
        post:
            summary: Logout from the current session.
            tags:
                - Session
            responses:
                '401':
                    $ref: '#/components/requests/sessions/responses/401'

                '200':
                    description: Current user had been logged out.
    /api/session/clear:
        post:
            summary: Delete the current user's account and log out.
            tags:
                - User
                - Session
            responses:
                '401':
                    $ref: '#/components/requests/sessions/responses/401'

                '200':
                    description: Current user had been deleted and logged out.
    /api/session/meta:
        get:
            summary: Returns metadata linked to the current user.
            tags:
                - Session
            responses:
                '401':
                    $ref: '#/components/requests/sessions/responses/401'

                '200':
                    description: Metadata linked to the current user
                    content:
                        $ref: '#/components/schemas/user_meta'
        post:
            summary: Update metadata linked to the current user.
            requestBody:
                content:
                    $ref: '#/components/schemas/user_meta'

            tags:
                - Session
            responses:
                '401':
                    $ref: '#/components/requests/sessions/responses/401'

                '200':
                    description: Metadata linked to the current user
                    content:
                        $ref: '#/components/schemas/user_meta'
...